+++ renku/ui/cli/utils/plugins.py
@@ -35,3 +35,10 @@ def get_supported_session_providers_names():
     from renku.core.plugin.session import get_supported_session_providers
 
     return [p.name for p in get_supported_session_providers()]
+
+
+def get_supported_hibernating_session_providers_names():
+    """Return names of session providers that support hibernation."""
+    from renku.core.plugin.session import get_supported_hibernating_session_providers
+
+    return [p.name for p in get_supported_hibernating_session_providers()]