+++ renku/ui/service/controllers/datasets_files_list.py
@@ -1,6 +1,5 @@
-#
-# Copyright 2020 - Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Copyright Swiss Data Science Center (SDSC). A partnership between
+# École Polytechnique Fédérale de Lausanne (EPFL) and
 # Eidgenössische Technische Hochschule Zürich (ETHZ).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -31,7 +30,7 @@ class DatasetsFilesListCtrl(ServiceCtrl, RenkuOperationMixin):
 
     def __init__(self, cache, user_data, request_data):
         """Construct a datasets files list controller."""
-        self.ctx = DatasetsFilesListCtrl.REQUEST_SERIALIZER.load(request_data)
+        self.ctx = self.REQUEST_SERIALIZER.load(request_data)
 
         super().__init__(cache, user_data, request_data)
 
@@ -42,10 +41,10 @@ class DatasetsFilesListCtrl(ServiceCtrl, RenkuOperationMixin):
 
     def renku_op(self):
         """Renku operation for the controller."""
-        result = list_files_command().build().execute(datasets=[self.ctx["name"]])
+        result = list_files_command().build().execute(datasets=[self.ctx["slug"]])
         return result.output
 
     def to_response(self):
         """Execute controller flow and serialize to service response."""
         self.ctx["files"] = self.execute_op()
-        return result_response(DatasetsFilesListCtrl.RESPONSE_SERIALIZER, self.ctx)
+        return result_response(self.RESPONSE_SERIALIZER, self.ctx)