+++ renku/ui/service/serializers/common.py
@@ -1,6 +1,5 @@
-#
-# Copyright 2020 - Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Copyright Swiss Data Science Center (SDSC). A partnership between
+# École Polytechnique Fédérale de Lausanne (EPFL) and
 # Eidgenössische Technische Hochschule Zürich (ETHZ).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -23,12 +22,21 @@ from marshmallow import Schema, fields, pre_load, validates
 
 from renku.ui.service.errors import UserRepoUrlInvalidError
 from renku.ui.service.serializers.rpc import JsonRPCResponse
+from renku.ui.service.utils import normalize_git_url
 
 
 class RemoteRepositoryBaseSchema(Schema):
     """Schema for tracking a remote repository."""
 
-    git_url = fields.String(metadata={"description": "Remote git repository url."})
+    git_url = fields.String(required=True, metadata={"description": "Remote git repository url."})
+
+    @pre_load
+    def normalize_url(self, data, **_):
+        """Remove ``.git`` extension from the git url."""
+        if "git_url" in data and data["git_url"]:
+            data["git_url"] = normalize_git_url(data["git_url"])
+
+        return data
 
     @validates("git_url")
     def validate_git_url(self, value):
@@ -48,7 +56,7 @@ class RemoteRepositorySchema(RemoteRepositoryBaseSchema):
     branch = fields.String(load_default=None, metadata={"description": "Remote git branch (or tag or commit SHA)."})
 
     @pre_load
-    def set_branch_from_ref(self, data, **kwargs):
+    def set_branch_from_ref(self, data, **_):
         """Set `branch` field from `ref` if present."""
         if "ref" in data and not data.get("branch"):
             # Backward compatibility: branch and ref were both used. Let's keep branch as the exposed field
@@ -57,6 +65,12 @@ class RemoteRepositorySchema(RemoteRepositoryBaseSchema):
             del data["ref"]
 
         return data
+
+
+class GitCommitSHA:
+    """Schema for a commit SHA."""
+
+    commit_sha = fields.String(load_default=None, metadata={"description": "Git commit SHA."})
 
 
 class AsyncSchema(Schema):
@@ -154,3 +168,9 @@ class ErrorResponse(Schema):
     userReference = fields.String()
     devReference = fields.String()
     sentry = fields.String()
+
+
+class GitUrlResponseMixin(Schema):
+    """Response containing a git url."""
+
+    git_url = fields.String(required=True, metadata={"description": "Remote git repository url."})