+++ renku/ui/service/serializers/config.py
@@ -1,6 +1,5 @@
-#
-# Copyright 2020 - Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Copyright Swiss Data Science Center (SDSC). A partnership between
+# École Polytechnique Fédérale de Lausanne (EPFL) and
 # Eidgenössische Technische Hochschule Zürich (ETHZ).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -18,11 +17,18 @@
 
 from marshmallow import Schema, fields
 
-from renku.ui.service.serializers.common import AsyncSchema, MigrateSchema, RemoteRepositorySchema, RenkuSyncSchema
+from renku.ui.service.serializers.common import (
+    AsyncSchema,
+    GitCommitSHA,
+    GitUrlResponseMixin,
+    MigrateSchema,
+    RemoteRepositorySchema,
+    RenkuSyncSchema,
+)
 from renku.ui.service.serializers.rpc import JsonRPCResponse
 
 
-class ConfigShowRequest(RemoteRepositorySchema):
+class ConfigShowRequest(RemoteRepositorySchema, GitCommitSHA):
     """Request schema for config show."""
 
 
@@ -32,7 +38,7 @@ class ConfigShowSchema(Schema):
     config = fields.Dict(metadata={"description": "Dictionary of configuration items."}, required=True)
 
 
-class ConfigShowResponse(ConfigShowSchema):
+class ConfigShowResponse(ConfigShowSchema, GitUrlResponseMixin):
     """Response schema for project config show."""
 
     default = fields.Dict(metadata={"description": "Dictionary of default configuration items."}, required=True)
@@ -48,7 +54,7 @@ class ConfigSetRequest(AsyncSchema, ConfigShowSchema, MigrateSchema, RemoteRepos
     """Request schema for config set."""
 
 
-class ConfigSetResponse(ConfigShowSchema, RenkuSyncSchema):
+class ConfigSetResponse(ConfigShowSchema, RenkuSyncSchema, GitUrlResponseMixin):
     """Response schema for project config set."""
 
     default = fields.Dict(metadata={"description": "Dictionary of default configuration items."})