+++ renku/ui/service/serializers/workflows.py
@@ -1,6 +1,5 @@
-#
-# Copyright 2020 - Swiss Data Science Center (SDSC)
-# A partnership between École Polytechnique Fédérale de Lausanne (EPFL) and
+# Copyright Swiss Data Science Center (SDSC). A partnership between
+# École Polytechnique Fédérale de Lausanne (EPFL) and
 # Eidgenössische Technische Hochschule Zürich (ETHZ).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -23,11 +22,11 @@ from marshmallow_oneofschema import OneOfSchema
 from renku.domain_model.dataset import DatasetCreatorsJson
 from renku.infrastructure.persistent import Persistent
 from renku.ui.cli.utils.plugins import get_supported_formats
-from renku.ui.service.serializers.common import RemoteRepositorySchema
+from renku.ui.service.serializers.common import GitCommitSHA, GitUrlResponseMixin, RemoteRepositorySchema
 from renku.ui.service.serializers.rpc import JsonRPCResponse
 
 
-class WorkflowPlansListRequest(RemoteRepositorySchema):
+class WorkflowPlansListRequest(RemoteRepositorySchema, GitCommitSHA):
     """Request schema for plan list view."""
 
 
@@ -74,7 +73,7 @@ class WorflowPlanEntryResponse(AbstractPlanResponse):
     children = fields.List(fields.String)
 
 
-class WorkflowPlansListResponse(Schema):
+class WorkflowPlansListResponse(GitUrlResponseMixin):
     """Response schema for plan list view."""
 
     plans = fields.List(fields.Nested(WorflowPlanEntryResponse), required=True)
@@ -86,7 +85,7 @@ class WorkflowPlansListResponseRPC(JsonRPCResponse):
     result = fields.Nested(WorkflowPlansListResponse)
 
 
-class WorkflowPlansShowRequest(RemoteRepositorySchema):
+class WorkflowPlansShowRequest(RemoteRepositorySchema, GitCommitSHA):
     """Request schema for plan show view."""
 
     plan_id = fields.String(required=True)
@@ -137,7 +136,7 @@ class ParameterSchema(ParameterBaseSchema):
     pass
 
 
-class PlanDetailsResponse(AbstractPlanResponse):
+class PlanDetailsResponse(AbstractPlanResponse, GitUrlResponseMixin):
     """Schema for Plan details."""
 
     last_executed = fields.DateTime()
@@ -191,7 +190,7 @@ class LinkSchema(Schema):
     sink_entries = fields.List(fields.Nested(ParameterTargetSchema), data_key="sinks")
 
 
-class CompositePlanDetailsResponse(AbstractPlanResponse):
+class CompositePlanDetailsResponse(AbstractPlanResponse, GitUrlResponseMixin):
     """Schema for Plan details."""
 
     steps = fields.List(fields.Nested(PlanReferenceSchema), data_key="plans")
@@ -223,7 +222,7 @@ WorkflowExportFormatEnum = Enum(  # type: ignore
 )
 
 
-class WorkflowPlansExportRequest(RemoteRepositorySchema):
+class WorkflowPlansExportRequest(RemoteRepositorySchema, GitCommitSHA):
     """Request schema for exporting a plan."""
 
     plan_id = fields.String(required=True)